+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH8E1PDDS9RW5582BX5B4SQX"
+        "01KH8VC3TCSKDA1SNSFYD5S699"
     )
     r"""The unique identifier of the data source"""