+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH8E1PD8Q07HT9XYPS3J9EBY"
+        "01KH8VC3T1BHJZ2X9ZBC1D029E"
     )
     r"""The unique identifier of the data source"""