+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH8E1PDBZM99BCJCJ2ST5G1J"
+        "01KH8VC3T4JMK4SDW95PA6BENE"
     )
     r"""The unique identifier of the data source"""