+++ packages/orq-rc/src/orq_ai_sdk/models/retrieveidentityop.py
@@ -103,7 +103,7 @@ class RetrieveIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-12T08:03:44.307Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-12T11:56:38.463Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")