+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -231,7 +231,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kh8e1nvk2jpt26a038hc3kc5"
+        "conv_01kh8vc38b8vnv6ekdp9r2h511"
     )
     r"""Unique conversation identifier with `conv_` prefix."""