+++ packages/orq-rc/RELEASES.md
@@ -5579,3 +5579,13 @@ Based on:
 - [python v4.3.0-rc.58] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.58] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.58 - packages/orq-rc
+
+## 2026-02-12 11:59:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.713.0 (2.821.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.59] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.59] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.59 - packages/orq-rc